fixed: 已经修复了角色和题库管理的相关bug

+++ backend/model/problem.go
@@ -2,8 +2,8 @@ package model
 
 type Problem struct {
 	ID              int64  `json:"problem_id"`
-	TimeLimit       int64  `json:"time_limit"`
-	MemoryLimit     int64  `json:"memory_limit"`
+	TimeLimit       int    `json:"time_limit"`
+	MemoryLimit     int    `json:"memory_limit"`
 	Name            string `json:"problem_name"`
 	Description     string `json:"description"`
 	TestCase        string `json:"test_case"`
